Module quartz::ir

Quartz IR

This module contains the data structures and functions that make up the QIR API. All IR is modeled in this module and this module alone.

This module also exposes the IR parsing/serialization APIs, as those are closely intertwined with the structure of the IR.

In memory, instructions are modeled using a doubly-linked list. This is due to the nature of an IR designed specifically for optimizations:

The data structure must be able to maintain an order (i.e. it has to be sequential)

The data structure must support fast random insertion/deletion

The data structure must support fast splicing

This effectively removes any possibility for clever data structures.

Structs

ArrayType

Models the [T; N] compound type.

BoolType

Models the bool fundamental type.

FloatType

Models the fN class of fundamental types.

IntType

Models the iN class of fundamental types.

PtrType

Models the ptr fundamental type.

StructType

Models the { T... } compound type.

TypeLayout

Models the layout requirements of a given type.

Enums

IEEEFloat

Maps the hardware representation of the floating-point types to enum variants.

Type

A single type in QIR.

Traits

IRType

Models the accessible operations on any given type.