Enum quartz::ir::Type

pub enum Type {
    Bool(BoolType),
    Int(IntType),
    Float(FloatType),
    Ptr(PtrType),
    Array(ArrayType),
    Struct(StructType),
}

A single type in QIR.

This is the public-facing type that is used whenever a

Variants

Bool(BoolType)

A boolean type.

See bool in the Reference

Int(IntType)

A raw integer type of some width.

See iN in the Reference

Float(FloatType)

An IEEE floating-point type of some width.

See fN in the Reference

Ptr(PtrType)

The ptr type, equivalent to C’s void*.

See ptr in the Reference

Array(ArrayType)

An array type of the form [T; N].

See [T; N] in the Reference

Struct(StructType)

A structure type of the form { T... }.

See { T... } in the Reference

Implementations

impl Type

pub const fn bool() -> Self

Creates a new bool type. Equivalent to Type::Bool(BoolType::new()).

Example

let ty = Type::bool();

assert!(ty.is_bool());

pub fn int(bit_width: u32) -> Self

Creates a new iN type. Equivalent to Type::Int(IntType::of_width(width)).

Example

let ty = Type::int(32);

assert!(ty.is_int());

Contracts

Pre-condition: bit_width is one of $ { 8, 16, 32, 64 }$

[8, 16, 32, 64].contains(& bit_width)

pub const fn float(ty: IEEEFloat) -> Self

Creates a type modeling fN from a given type of float.

Example

let ty = Type::float(IEEEFloat::Single);

assert_eq!(ty.is_float(), true)

pub const fn ptr() -> Self

Creates a new ptr type. Equivalent to Type::Ptr(PtrType::new()).

Example

let ty = Type::ptr();

assert_eq!(ty.is_ptr(), true);

pub fn array(ty: Type, len: u64) -> Self

Creates a new [T; N] type. Equivalent to Type::Array(ArrayType::new(ty, len)).

Example

let ty = Type::array(Type::i32(), 16);

assert_eq!(ty.is_array(), true);

pub fn structure<T: IntoIterator<Item = Type>>(members: T) -> Self

Creates a new { T... } type. Equivalent to Type::Struct(StructType::new(members)).

Example

let ty = Type::structure([Type::f64(), Type::f64()]);

assert_eq!(ty.is_struct(), true);

pub const fn i8() -> Self

Creates a Type that models i8. Equivalent to Type::int(8).

Example

let ty1 = Type::i8();
let ty2 = Type::int(8);

assert_eq!(ty1, ty2);

pub const fn i16() -> Self

Creates a Type that models i16. Equivalent to Type::int(16).

Example

let ty1 = Type::i8();
let ty2 = Type::int(8);

assert_eq!(ty1, ty2);

pub const fn i32() -> Self

Creates a Type that models i32. Equivalent to Type::int(32).

Example

let ty1 = Type::i8();
let ty2 = Type::int(8);

assert_eq!(ty1, ty2);

pub const fn i64() -> Self

Creates a Type that models i64. Equivalent to Type::int(64).

Example

let ty1 = Type::i8();
let ty2 = Type::int(8);

assert_eq!(ty1, ty2);

pub const fn f32() -> Self

Creates a Type that models f32. Equivalent to Type::float(FloatType::Single).

Example

let ty1 = Type::f32();
let ty2 = Type::float(IEEEFloat::Single);

assert_eq!(ty1, ty2);

pub const fn f64() -> Self

Creates a Type that models f64. Equivalent to Type::float(FloatType::Double).

Example

let ty1 = Type::f64();
let ty2 = Type::float(IEEEFloat::Double);

assert_eq!(ty1, ty2);

pub fn empty_struct() -> Self

Creates the empty { } type. Equivalent to Type::Struct(StructType::empty()).

Example

let ty = Type::empty_struct();

assert_eq!(ty.is_struct(), true);

pub fn is_bool(&self) -> bool

Checks if self is currently a Bool.

let ty = Type::bool();
assert_eq!(ty.is_bool(), true);
assert_eq!(ty.is_ptr(), false);

pub fn unwrap_bool(self) -> BoolType

Checks if the current variant is Bool. If it is, it returns the data for that variant.

Panics

Panics if the current variant is not Bool.

Examples

let ty = Type::bool();
let data: BoolType = ty.unwrap_bool();

let ty = Type::i32();
let data: BoolType = ty.unwrap_bool(); // panics!

pub fn expect_bool(self, message: &str) -> BoolType

Checks if the current variant is Bool. If it is, it returns the data for that variant.

Panics

Panics with message if the current variant is not Bool.

Examples

let ty = Type::bool();
let data: BoolType = ty.expect_bool("should be the right type");

let ty = Type::i32();
let data: BoolType = ty.expect_bool("should be the right type"); // panics!

pub fn is_int(&self) -> bool

Checks if self is currently a Int.

let ty = Type::i64();
assert_eq!(ty.is_int(), true);
assert_eq!(ty.is_bool(), false);

pub fn unwrap_int(self) -> IntType

Checks if the current variant is Int. If it is, it returns the data for that variant.

Panics

Panics if the current variant is not Int.

Examples

let ty = Type::i64();
let data: IntType = ty.unwrap_int();

let ty = Type::ptr();
let data: IntType = ty.unwrap_int(); // panics!

pub fn expect_int(self, message: &str) -> IntType

Checks if the current variant is Int. If it is, it returns the data for that variant.

Panics

Panics with message if the current variant is not Int.

Examples

let ty = Type::i64();
let data: IntType = ty.expect_int("should be the right type");

let ty = Type::ptr();
let data: IntType = ty.expect_int("should be the right type"); // panics!

pub fn is_float(&self) -> bool

Checks if self is currently a Float.

let ty = Type::f64();
assert_eq!(ty.is_float(), true);
assert_eq!(ty.is_int(), false);

pub fn unwrap_float(self) -> FloatType

Checks if the current variant is Float. If it is, it returns the data for that variant.

Panics

Panics if the current variant is not Float.

Examples

let ty = Type::f64();
let data: FloatType = ty.unwrap_float();

let ty = Type::i32();
let data: FloatType = ty.unwrap_float(); // panics!

pub fn expect_float(self, message: &str) -> FloatType

Checks if the current variant is Float. If it is, it returns the data for that variant.

Panics

Panics with message if the current variant is not Float.

Examples

let ty = Type::f64();
let data: FloatType = ty.expect_float("should be the right type");

let ty = Type::i32();
let data: FloatType = ty.expect_float("should be the right type"); // panics!

pub fn is_ptr(&self) -> bool

Checks if self is currently a Ptr.

let ty = Type::ptr();
assert_eq!(ty.is_ptr(), true);
assert_eq!(ty.is_array(), false);

pub fn unwrap_ptr(self) -> PtrType

Checks if the current variant is Ptr. If it is, it returns the data for that variant.

Panics

Panics if the current variant is not Ptr.

Examples

let ty = Type::ptr();
let data: PtrType = ty.unwrap_ptr();

let ty = Type::bool();
let data: PtrType = ty.unwrap_ptr(); // panics!

pub fn expect_ptr(self, message: &str) -> PtrType

Checks if the current variant is Ptr. If it is, it returns the data for that variant.

Panics

Panics with message if the current variant is not Ptr.

Examples

let ty = Type::ptr();
let data: PtrType = ty.expect_ptr("should be the right type");

let ty = Type::bool();
let data: PtrType = ty.expect_ptr("should be the right type"); // panics!

pub fn is_array(&self) -> bool

Checks if self is currently a Array.

let ty = Type::array(Type::ptr(), 512);
assert_eq!(ty.is_array(), true);
assert_eq!(ty.is_ptr(), false);

pub fn unwrap_array(self) -> ArrayType

Checks if the current variant is Array. If it is, it returns the data for that variant.

Panics

Panics if the current variant is not Array.

Examples

let ty = Type::array(Type::ptr(), 512);
let data: ArrayType = ty.unwrap_array();

let ty = Type::f32();
let data: ArrayType = ty.unwrap_array(); // panics!

pub fn expect_array(self, message: &str) -> ArrayType

Checks if the current variant is Array. If it is, it returns the data for that variant.

Panics

Panics with message if the current variant is not Array.

Examples

let ty = Type::array(Type::ptr(), 512);
let data: ArrayType = ty.expect_array("should be the right type");

let ty = Type::f32();
let data: ArrayType = ty.expect_array("should be the right type"); // panics!

pub fn is_struct(&self) -> bool

Checks if self is currently a Struct.

let ty = Type::structure([Type::ptr(), Type::i64(), Type::i64()]);
assert_eq!(ty.is_struct(), true);
assert_eq!(ty.is_int(), false);

pub fn unwrap_struct(self) -> StructType

Checks if the current variant is Struct. If it is, it returns the data for that variant.

Panics

Panics if the current variant is not Struct.

Examples

let ty = Type::structure([Type::ptr(), Type::i64(), Type::i64()]);
let data: StructType = ty.unwrap_struct();

let ty = Type::ptr();
let data: StructType = ty.unwrap_struct(); // panics!

pub fn expect_struct(self, message: &str) -> StructType

Checks if the current variant is Struct. If it is, it returns the data for that variant.

Panics

Panics with message if the current variant is not Struct.

Examples

let ty = Type::structure([Type::ptr(), Type::i64(), Type::i64()]);
let data: StructType = ty.expect_struct("should be the right type");

let ty = Type::ptr();
let data: StructType = ty.expect_struct("should be the right type"); // panics!