Trait quartz::ir::IRType

pub trait IRType {
    fn size_req(&self) -> Option<u64> { ... }
    fn align_req(&self) -> Option<u64> { ... }
    fn layout(&self) -> Option<TypeLayout> { ... }
}

Models the accessible operations on any given type.

Note that any type includes odd types, such as opaque structures, zero-length arrays, empty structures, etc.

fn check_if_16_bytes_large(ty: &dyn IRType) -> bool {
    match ty.layout() {
        Some(layout) => layout.size() == 16,
        None => false
    }
}

Provided Methods

fn size_req(&self) -> Option<u64>

Gets the exact size (in bytes) of the types. If the type is zero-sized, None is returned. Otherwise, Some(size_in_bytes) is returned.

This size can be considered as the minimum amount of storage required to properly store the given type.

Alignment is either the same or smaller than this size.

Unchecked Contracts

ret >= self.align_req(): The return value of [align_req] never exceeds the return value of this method.

ret.is_none() == self.align_req().is_none(): If the return value is None, so will be the return value of [align_req].

Contracts

Post-condition: ret is never Some(0)

ret != Some(0)

fn align_req(&self) -> Option<u64>

Gets the size (in bytes) of the types. If the type is zero-sized, None is returned. Otherwise, Some(size_in_bytes) is returned.

Unchecked Contracts

ret <= self.size_req(): The return value of [size_req] is always the same as or greater than the return value of this method.

ret.is_none() == self.size_req().is_none(): If the return value is None, so will be the return value of [size_req].

Contracts

Post-condition: ret is never Some(0)

ret != Some(0)

Post-condition: alignment is always a power of 2

ret.is_none() || ret.unwrap().is_power_of_two()

fn layout(&self) -> Option<TypeLayout>

The preferred way of getting both size and alignment data for a type.

This better codifies the unchecked contracts of [size_req] and [align_req], as if both would be None this returns None.

If this returns None, the type is zero-sized and therefore has no layout requirements. Otherwise, the [size] and [align] functions return the requirements.

Example

let ty = Type::f64();
let layout = ty.layout().unwrap();

assert_eq!(layout.size(), 8);
assert_eq!(layout.align(), 8);

Contracts

Post-condition: ret is never Some(0)

ret != Some(TypeLayout { size : 0, align : 0 })